Refactor favorites page and PhotoResults - WIP
- Extract the react logic from the favorites page and PhotoResults into useFavorites and usePhotos hook
- Create imageHelper
